refactor(api): move shared prompt primitives to a shared module
Extract the version-agnostic enrichment prompt primitives (MISSION_DATA_TAG,
ENRICHMENT_SCHEMA, NON_ENRICHABLE_VALUE_KEYS, buildFilteredTaxonomyBlock,
buildTaxonomyGuidanceBlock, buildUserMessage) into a neutral prompts/shared.ts
module.

v1, v2 and v5 now consume them from shared.ts instead of redefining them; v2/v1
re-export what v3/v4 still import. v1 keeps its own legacy buildUserMessage
variant (different format). No prompt version depends on another version, and
behavior is unchanged (rendered prompts and user messages are identical).
